Remove get_workflow_actions from bika_listing, being used in analysestransposed though
